requirements.txt
yt_dlp
opencv-python
scikit-learn
numpy
python-dotenv
openai
pydantic
pillow
matplotlib